requirements.txt
boto3==1.34.0
python-dateutil==2.8.2
Pillow==10.1.0
pytest==7.4.3
pytest-mock==3.12.0
moto==4.2.14
requests==2.31.0

